% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plapi.R
\name{Client}
\alias{Client}
\title{plapi constructor}
\usage{
Client(api_url, api_user, api_password, api_key, api_hash)
}
\arguments{
\item{api_url}{url of API requested}

\item{api_user}{Username}

\item{api_password}{Password}

\item{api_key}{API Key}

\item{api_hash}{API Hash}
}
\description{
This function creates a client for the PLAPIClient API
}
